pyproject.toml
[project]
name = "semantic-kernel-first-play"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "openai>=1.77.0",
    "python-dotenv>=1.1.0",
    "semantic-kernel>=1.29.0",
    "typing>=3.10.0.0",
]
